requirements.txt
# Linters
flake8>=3.5.0,<4.0
autopep8>=1.3,<2.0
pycodestyle>=2.3.1,<3.0

# Tests and utilities
PyYAML>=3.12,<4.0

# Tests
mock>=2.0.0,<3.0
salt==2017.7.2
